+++ src/main/java/ru/yandex/practicum/filmorate/storage/UserStorage.java
@@ -3,11 +3,11 @@ package ru.yandex.practicum.filmorate.storage;
 import org.springframework.web.bind.annotation.RequestBody;
 import ru.yandex.practicum.filmorate.model.User;
 
-interface UserStorage {
+interface UserStorage  {
 
     User createUser(User user);
 
-    void deleteUser(Long id);
+    void deleteUser(int id);
 
     User updateUser(User user);
 }
